refactor: extract CommandDialog
