+++ src/web/src/__tests__/components/WSEditorCommandArgumentsContent.test.tsx
@@ -1,11 +1,8 @@
 import { describe, it, expect, vi, beforeEach } from "vitest";
 import { fireEvent, screen } from "@testing-library/react";
 import WSEditorCommandArgumentsContent from "../../views/workspace/WSEditorCommandArgumentsContent/WSEditorCommandArgumentsContent";
-import type {
-  CMDArg,
-  ClsArgDefinitionMap,
-} from "../../views/workspace/WSEditorCommandArgumentsContent/WSEditorCommandArgumentsContent";
 import { render } from "../test-utils";
+import { ClsArgDefinitionMap, CMDArg } from "../../views/workspace/WSEditorCommandArgumentsContent";
 
 vi.mock("../../services/commandApi");
 vi.mock("../../services/errorHandlerApi");
