+++ src/web/src/views/workspace/WSEditorCommandContent/index.ts
@@ -1,2 +1,3 @@
-export { default, DecodeResponseCommand } from "./WSEditorCommandContent";
+export { default } from "./WSEditorCommandContent";
+export { DecodeResponseCommand } from "../utils/decodeResponseCommand";
 export type { Command, Resource, ResponseCommand, ResponseCommands, Example, Plane } from "../interfaces";
